% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CODEXindex.R
\name{codex}
\alias{codex}
\title{The CODEX Index: Short and Medium-term Prognosis in Patients Hospitalized for COPD Exacerbations}
\source{
\url{https://www.ncbi.nlm.nih.gov/pubmed/24077342}
}
\usage{
codex(age, charlson, FEV1, mMRC, exacerbation)
}
\arguments{
\item{age}{a number for age}

\item{charlson}{a number for charlson index}

\item{FEV1}{a number for forced expiratory volume in one second (FEV1) expressed as percentage}

\item{mMRC}{a number for the dyspnea scale of modified Medical Research Council (mMRC)}

\item{exacerbation}{a number for severe exacerbations of COPD during the previous year (hospitalization or ED consultations)}
}
\description{
The CODEX Index: Short and Medium-term Prognosis in Patients Hospitalized for COPD Exacerbations
}
\examples{
codex (age = 40,charlson = 8 , FEV1 = 40, mMRC = 3, exacerbation = 2)
codex (age = 55,charlson = 4 , FEV1 = 40, mMRC = 3, exacerbation = 2)

}
